llow dynamic selection of the method of comm w/vbox.
The vmfest provider now can cake a new parameter `:vbox-comm` to
select the method of communication of Pallet with the VirtualBox
subsystem. The options are :xpcom for XPCOM, and :ws for Web Services.
It defaults to :xpcom.

There is no longer the need to add either vboxjxws or vboxjxpcom jars
to the classpath prior to using pallet-vmfest, in order to select the
communication method. Pallet-vmfest contains both jars and one of them
will be injected in the classpath dynamically depending on the value
of `:vbox-comm`.

NOTE: Selecting the communication methods with VirtualBox can only be
done once per JVM session.