Ignore stopped/inaccessible VMs
If a VM is misconfigured, or stopped, pallet cannot manage it. From
now on these VMs will be effectively ignored. For example, if you have
a `vm-test-0` VM that is stopped, and you converge to 1 the same
group, you will get another VM in this group. Converging back to 0
will remove this new vm, leaving the stopped `vm-test-0` alone.

Similarly, all VMs created by pallet are in the `vmfest` group in
VirtualBox, and pallet will not _see_ any VM outside of this group.
The VMs outside the pallet group were already ignored, since they
don't have any group tag, but now they will not even be
considered (e.g. not listed in pallet.repl/show-nodes).